requirements.txt
Flask==3.1.3
boto3==1.34.117
pymongo==4.8.0
python-dotenv==1.1.1
gunicorn==23.0.0
flask-cors==5.0.1